chore(aptos-proof-inclusion): fixing xclippy
