--- crates/merkle-inclusion/tests/lib.rs
@@ -1,2 +1,0 @@
-pub(crate) mod gadget;
-pub(crate) mod utils;
